refactor: toggle 메서드 사용 및 script 모듈화
